Access Time Diagnosis
Adding more options to see the access time and find out if it's doing what
you expect. The list of candidates now contains the access date by default
and you can also turn on `-v` or `--verbose` to get information about how
the access date for each entry was calculated.

[fixes #10]